Remove feature for passing extra arguments to buildah-from command
The reason for this feature was to create a container
without bind-mounting /etc/resolv.conf from the host,
using the option '--dns none'. This way a playbook being
executed inside the container could override that file,
but obviously this also breaks name lookups.

An easier alternative would be to create one task in
arch_container role to copy the desired file into
container's mount point.

For the moment it's fine to leave default resolv.conf.